feat: add pack-owned processor contracts
