+++ src/openclaw_pipeline/packs/research_tech/handlers.py
@@ -187,7 +187,7 @@ def build_stage_handlers(pack_name: str = "research-tech") -> list[StageHandlerS
             handler_kind="focused_action",
             runtime_adapter="focused_action",
             action_kind="deep_dive_workflow",
-            entrypoint="openclaw_pipeline.truth_api:_run_deep_dive_workflow_action",
+            entrypoint="openclaw_pipeline.packs.research_tech.focused_actions:run_deep_dive_workflow_action",
             description="Create a deep dive for one processed source note",
             target_mode="single_note",
             safe_to_run=True,
@@ -200,7 +200,7 @@ def build_stage_handlers(pack_name: str = "research-tech") -> list[StageHandlerS
             handler_kind="focused_action",
             runtime_adapter="focused_action",
             action_kind="object_extraction_workflow",
-            entrypoint="openclaw_pipeline.truth_api:_run_object_extraction_workflow_action",
+            entrypoint="openclaw_pipeline.packs.research_tech.focused_actions:run_object_extraction_workflow_action",
             description="Extract evergreen objects for one deep dive",
             target_mode="single_note",
             safe_to_run=True,
