+++ src/openclaw_pipeline/packs/research_tech/processor_contracts.py
@@ -1,0 +1,85 @@
+from __future__ import annotations
+
+from ..base import ProcessorContractSpec
+
+
+def build_processor_contracts(pack_name: str = "research-tech") -> list[ProcessorContractSpec]:
+    return [
+        ProcessorContractSpec(
+            name="articles",
+            pack=pack_name,
+            stage="articles",
+            mode="llm_structured",
+            inputs=("source_note",),
+            outputs=("deep_dive",),
+            quality_hooks=("quality",),
+            entrypoint="openclaw_pipeline.workflow_handlers:run_pipeline_articles",
+            description="Interpret one or more processed source notes into deep dives.",
+        ),
+        ProcessorContractSpec(
+            name="quality",
+            pack=pack_name,
+            stage="quality",
+            mode="evaluation",
+            inputs=("deep_dive",),
+            outputs=("quality_report",),
+            quality_hooks=(),
+            entrypoint="openclaw_pipeline.workflow_handlers:run_pipeline_quality",
+            description="Evaluate deep dive quality and emit review signals.",
+        ),
+        ProcessorContractSpec(
+            name="absorb",
+            pack=pack_name,
+            stage="absorb",
+            mode="hybrid",
+            inputs=("deep_dive",),
+            outputs=("evergreen_object", "relation", "compiled_summary"),
+            quality_hooks=("quality",),
+            entrypoint="openclaw_pipeline.workflow_handlers:run_pipeline_absorb",
+            description="Project qualified deep dives into evergreen truth artifacts.",
+        ),
+        ProcessorContractSpec(
+            name="moc",
+            pack=pack_name,
+            stage="moc",
+            mode="rule_based",
+            inputs=("evergreen_object", "relation"),
+            outputs=("atlas_page",),
+            quality_hooks=(),
+            entrypoint="openclaw_pipeline.workflow_handlers:run_pipeline_moc",
+            description="Refresh atlas/MOC navigation from projected knowledge objects.",
+        ),
+        ProcessorContractSpec(
+            name="knowledge_index",
+            pack=pack_name,
+            stage="knowledge_index",
+            mode="projection_rebuild",
+            inputs=("vault_state",),
+            outputs=("truth_projection", "graph_projection"),
+            quality_hooks=(),
+            entrypoint="openclaw_pipeline.workflow_handlers:run_pipeline_knowledge_index",
+            description="Rebuild the derived truth and graph projections.",
+        ),
+        ProcessorContractSpec(
+            name="deep_dive_workflow",
+            pack=pack_name,
+            action_kind="deep_dive_workflow",
+            mode="llm_structured",
+            inputs=("source_note",),
+            outputs=("deep_dive",),
+            quality_hooks=("quality",),
+            entrypoint="openclaw_pipeline.packs.research_tech.focused_actions:run_deep_dive_workflow_action",
+            description="Focused single-note deep dive generation action.",
+        ),
+        ProcessorContractSpec(
+            name="object_extraction_workflow",
+            pack=pack_name,
+            action_kind="object_extraction_workflow",
+            mode="hybrid",
+            inputs=("deep_dive",),
+            outputs=("evergreen_object", "relation", "compiled_summary"),
+            quality_hooks=("quality",),
+            entrypoint="openclaw_pipeline.packs.research_tech.focused_actions:run_object_extraction_workflow_action",
+            description="Focused single-note evergreen extraction action.",
+        ),
+    ]
