+++ tests/test_doctor_command.py
@@ -20,8 +20,16 @@ def test_doctor_command_reports_primary_and_compatibility_roles(capsys):
     assert payload["docs"]["verify"]["exists"] is True
     assert payload["contracts"]["declared"]["truth_projection"]["pack"] == "research-tech"
     assert any(
+        item["name"] == "articles" and item["mode"] == "llm_structured"
+        for item in payload["contracts"]["declared"]["processor_contracts"]
+    )
+    assert any(
         item["action_kind"] == "deep_dive_workflow"
         for item in payload["contracts"]["effective"]["stage_handlers"]
+    )
+    assert any(
+        item["action_kind"] == "deep_dive_workflow"
+        for item in payload["contracts"]["effective"]["processor_contracts"]
     )
     assert {
         item["surface_kind"] for item in payload["contracts"]["effective"]["observation_surfaces"]
@@ -42,6 +50,11 @@ def test_doctor_command_reports_compatibility_pack_metadata(capsys):
     assert payload["docs"]["verify"]["exists"] is False
     assert payload["contracts"]["declared"]["truth_projection"] is None
     assert payload["contracts"]["effective"]["truth_projection"]["pack"] == "research-tech"
+    assert payload["contracts"]["declared"]["processor_contracts"] == []
+    assert any(
+        item["stage"] == "articles" and item["pack"] == "research-tech"
+        for item in payload["contracts"]["effective"]["processor_contracts"]
+    )
     assert any(
         item["runtime_adapter"] == "focused_action"
         for item in payload["contracts"]["effective"]["stage_handlers"]
